# admin-frontend/src/main/frontend/openapi/admin-api.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "contract"
  },
  "paths" : {
...
    "/admin/storage/cleanup" : {
      "post" : {
        "tags" : [ "storage" ],
        "operationId" : "triggerCleanup",
        "parameters" : [ {
          "name" : "repoId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CacheCleanupResult"
                }
              }
            }
          }
        },
        "security" : [ {
          "basicAuth" : [ ]
        } ]
      }
    },
...
    "/admin/storage/usage" : {
      "get" : {
        "tags" : [ "storage" ],
        "operationId" : "getStorageUsage",
        "parameters" : [ {
          "name" : "repoId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "exact",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        }, {
          "name" : "bucket",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StorageUsageInfo"
                }
              }
            }
          }
        },
        "security" : [ {
          "basicAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ErrorMessage" : {
        "properties" : {
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "details" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "exception" : {
            "type" : "object",
            "properties" : {
              "cause" : {
                "type" : "object",
                "properties" : {
                  "stackTrace" : {
                    "type" : "array",
                    "items" : {
                      "type" : "object",
                      "properties" : {
                        "classLoaderName" : {
                          "type" : "string"
                        },
                        "moduleName" : {
                          "type" : "string"
                        },
                        "moduleVersion" : {
                          "type" : "string"
                        },
                        "methodName" : {
                          "type" : "string"
                        },
                        "fileName" : {
                          "type" : "string"
                        },
                        "lineNumber" : {
                          "type" : "integer",
                          "format" : "int32"
                        },
                        "className" : {
                          "type" : "string"
                        },
                        "nativeMethod" : {
                          "type" : "boolean"
                        }
                      }
                    }
                  },
                  "message" : {
                    "type" : "string"
                  },
                  "localizedMessage" : {
                    "type" : "string"
                  }
                }
              },
              "stackTrace" : {
                "type" : "array",
                "items" : {
                  "type" : "object",
                  "properties" : {
                    "classLoaderName" : {
                      "type" : "string"
                    },
                    "moduleName" : {
                      "type" : "string"
                    },
                    "moduleVersion" : {
                      "type" : "string"
                    },
                    "methodName" : {
                      "type" : "string"
                    },
                    "fileName" : {
                      "type" : "string"
                    },
                    "lineNumber" : {
                      "type" : "integer",
                      "format" : "int32"
                    },
                    "className" : {
                      "type" : "string"
                    },
                    "nativeMethod" : {
                      "type" : "boolean"
                    }
                  }
                }
              },
              "message" : {
                "type" : "string"
              },
              "suppressed" : {
                "type" : "array",
                "items" : {
                  "type" : "object",
                  "properties" : {
                    "stackTrace" : {
                      "type" : "array",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "classLoaderName" : {
                            "type" : "string"
                          },
                          "moduleName" : {
                            "type" : "string"
                          },
                          "moduleVersion" : {
                            "type" : "string"
                          },
                          "methodName" : {
                            "type" : "string"
                          },
                          "fileName" : {
                            "type" : "string"
                          },
                          "lineNumber" : {
                            "type" : "integer",
                            "format" : "int32"
                          },
                          "className" : {
                            "type" : "string"
                          },
                          "nativeMethod" : {
                            "type" : "boolean"
                          }
                        }
                      }
                    },
                    "message" : {
                      "type" : "string"
                    },
                    "localizedMessage" : {
                      "type" : "string"
                    }
                  }
                }
              },
              "localizedMessage" : {
                "type" : "string"
              }
            },
            "writeOnly" : true
          },
          "userMessage" : {
            "type" : "string"
          },
          "logLevel" : {
            "type" : "string"
          },
          "stacktrace" : {
            "type" : "string"
          }
        },
        "required" : [ "details", "logLevel" ]
      },
...
      "CacheCleanupResult" : {
        "properties" : {
          "scopesChecked" : {
            "type" : "integer",
            "format" : "int32"
          },
          "scopesCleaned" : {
            "type" : "integer",
            "format" : "int32"
          },
          "deletedEntries" : {
            "type" : "integer",
            "format" : "int32"
          },
          "freedBytes" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "deletedEntries", "freedBytes", "scopesChecked", "scopesCleaned" ]
      },
...
      "BucketUsageInfo" : {
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "size" : {
            "type" : "integer",
            "format" : "int64"
          },
          "quota" : {
            "type" : "integer",
            "format" : "int64"
          },
          "usedPercent" : {
            "type" : "number",
            "format" : "double"
          },
          "enforced" : {
            "type" : "boolean"
          },
          "measured" : {
            "type" : "boolean"
          },
          "kind" : {
            "type" : "string",
            "enum" : [ "BUCKET", "LIBRARY_CACHE" ]
          }
        },
        "required" : [ "enforced", "kind", "measured", "name", "size" ]
      },
      "StorageUsageInfo" : {
        "properties" : {
          "repoId" : {
            "type" : "string"
          },
          "totalSize" : {
            "type" : "integer",
            "format" : "int64"
          },
          "quota" : {
            "type" : "integer",
            "format" : "int64"
          },
          "usedPercent" : {
            "type" : "number",
            "format" : "double"
          },
          "exact" : {
            "type" : "boolean"
          },
          "buckets" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/BucketUsageInfo"
            }
          }
        },
        "required" : [ "buckets", "exact", "repoId", "totalSize" ]
      },
...
    },
    "securitySchemes" : {
      "basicAuth" : {
        "type" : "http",
        "scheme" : "basic"
      },
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}